refactor/data 폴더 request, response로 분리
