+++ app/src/main/java/com/example/yourun/model/data/request/RunningResultRequest.kt
@@ -1,4 +1,4 @@
-package com.example.yourun.model.data
+package com.example.yourun.model.data.request
 
 data class RunningResultRequest(
     val targetTime: Int,
