+++ app/src/main/java/com/example/yourun/model/data/request/SignupRequest.kt
@@ -1,4 +1,4 @@
-package com.example.yourun.model.data
+package com.example.yourun.model.data.request
 
 data class SignUpRequest(
     val email: String,
