+++ app/src/main/java/com/example/yourun/model/data/response/ApiResponseBoolean.kt
@@ -1,4 +1,4 @@
-package com.example.yourun.model.data
+package com.example.yourun.model.data.response
 
 data class ApiResponseBoolean(
     val status: Int,
