+++ app/src/main/java/com/example/yourun/model/data/response/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.example.yourun.model.data
+package com.example.yourun.model.data.response
 
 import com.google.gson.annotations.SerializedName
