+++ app/src/main/java/com/example/yourun/model/data/response/UserSoloChallengeInfo.kt
@@ -1,4 +1,4 @@
-package com.example.yourun.model.data
+package com.example.yourun.model.data.response
 
 data class UserSoloChallengeInfo(
     val challengeId: Long,
