--- app/src/main/java/com/example/yourun/model/network/ApiClientNoAuth.kt
@@ -1,37 +1,0 @@
-package com.example.yourun.model.network
-
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-object ApiClientNoAuth {
-    private val client = OkHttpClient.Builder()
-        .connectTimeout(30, TimeUnit.SECONDS)
-        .readTimeout(30, TimeUnit.SECONDS)
-        .writeTimeout(30, TimeUnit.SECONDS)
-        .apply {
-            val logging = HttpLoggingInterceptor()
-            logging.setLevel(HttpLoggingInterceptor.Level.BODY)
-            addInterceptor(logging)
-        }
-        .build()
-
-    fun getApiService(): ApiService {
-        val gson = GsonBuilder()
-            .serializeNulls()
-            .setLenient()
-            .create()
-
-        return Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(client)
-            .addConverterFactory(GsonConverterFactory.create(gson))
-            .build()
-            .create(ApiService::class.java)
-    }
-
-}
