+++ app/src/main/java/com/example/yourun/model/network/RunningApiService.kt
@@ -1,7 +1,7 @@
 package com.example.yourun.model.network
 
-import com.example.yourun.model.data.RunningResultRequest
-import com.example.yourun.model.data.RunningResultResponse
+import com.example.yourun.model.data.request.RunningResultRequest
+import com.example.yourun.model.data.response.RunningResultResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.Header
