+++ app/src/main/java/com/example/yourun/model/repository/LoginRepository.kt
@@ -1,8 +1,8 @@
 package com.example.yourun.model.repository
 
 import com.example.yourun.model.network.ApiService
-import com.example.yourun.model.data.LoginRequest
-import com.example.yourun.model.data.LoginResponse
+import com.example.yourun.model.data.request.LoginRequest
+import com.example.yourun.model.data.response.LoginResponse
 import retrofit2.Response
 
 class LoginRepository(private val apiService: ApiService) {
