+++ app/src/main/java/com/example/yourun/view/activities/CalendarActivity.kt
@@ -14,10 +14,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.lifecycleScope
 import com.example.yourun.R
-import com.example.yourun.model.data.RunningStatsResponse
+import com.example.yourun.model.data.response.RunningStatsResponse
 import com.example.yourun.model.network.ApiClient
-import com.example.yourun.model.network.ApiResponse
-import com.google.android.gms.common.api.Response
 import kotlinx.coroutines.launch
 import java.time.LocalDate
 import java.time.YearMonth
